Commit includes: add Feed component
